The GetSize was calculating the size each time, though it only changes in a few calls. I added a new field to keep track of the current size so that GetSize just returns the current size.

This reduces the encoding time by about 10% for C++.

```
-----------------------------------------------------------------------
Benchmark                             Time             CPU   Iterations
-----------------------------------------------------------------------
BM_Flatbuffers_Encode_mean          216 ns          216 ns            5
BM_Flatbuffers_Encode_median        216 ns          218 ns            5
BM_Flatbuffers_Encode_stddev       1.48 ns         2.48 ns            5
BM_Flatbuffers_Encode_cv           0.69 %          1.15 %             5
```

```
-----------------------------------------------------------------------
Benchmark                             Time             CPU   Iterations
-----------------------------------------------------------------------
BM_Flatbuffers_Encode_mean          189 ns          188 ns            5
BM_Flatbuffers_Encode_median        189 ns          188 ns            5
BM_Flatbuffers_Encode_stddev      0.996 ns         2.96 ns            5
BM_Flatbuffers_Encode_cv           0.53 %          1.57 %             5
```